requirements.txt
streamlit==1.29.0
pandas==2.2.0
python-dotenv==1.0.0
evolutionapi==0.0.9
openpyxl==3.1.2
numpy==1.26.3
datetime==5.3